Update the first page and add a .gitignore for _site/ folder
So this was more than what I intended to put in, but I updated
the text on the first page and also removed the _site/ folder, where
all the autogenerated stuff shows up.

Everything works fine with it in there, it's just that committing
all those artefacts isn't necessary for us at all.

Also removed some leftovers from the config.yml file.