Add netlify.toml file with redirect rule
The idea here is to put a redirect in so that we never go to
the index page. The index page isn't editable using the CMS, so
we are better off just hiding it.

We aren't really able to put extra stuff into the _site directory,
so I used the .toml file option.

Source:
- https://docs.netlify.com/routing/redirects/#syntax-for-the-netlify-configuration-file