Add some word breaks to the continuing docs page
Also make text inside tables smaller on mobile.

Some of the content on the continuing docs page (en and fr) was pushing
the layout off of very small screens.

The reason was that we had some long-ish filenames without any linebreaks
between them. Added some line breaks and made text in tables smaller
so that the content all fits on the screen.